fix ranking videos

+++ app/Http/Controllers/GenerickeyController.php
@@ -159,7 +159,11 @@ class GenerickeyController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id): View{}
+    public function show($id): RedirectResponse
+    {
+        return redirect()->route($this->url_path.'.index');
+    }
+
     
     /**
      * Show the form for editing the specified resource.
@@ -169,6 +173,7 @@ class GenerickeyController extends Controller
      */
     public function edit($id): View{
         $m_data = Generickey::find($id);
+        abort_if(!$m_data, 404);
         $m_data_descriptions = DB::table($this->description_table_name)->select()->where('key_id',$id)->get()->keyBy('language_id');
         $data = array();
         $data['module_title_singular'] = $this->module_title_singular;
@@ -205,6 +210,7 @@ class GenerickeyController extends Controller
             $input['status']=0;
         }
         $m_data = Generickey::find($id);
+        abort_if(!$m_data, 404);
         $m_data->update($input);
 
         foreach ($site_languages as $language) {
@@ -232,7 +238,9 @@ class GenerickeyController extends Controller
      */
     public function destroy($id): RedirectResponse
     {
-        Generickey::find($id)->delete();
+        $record = Generickey::find($id);
+        abort_if(!$record, 404);
+        $record->delete();
         DB::table($this->description_table_name)->where('key_id',$id)->delete();
         return redirect()->route($this->url_path.'.index')->with('success',$this->module_title_singular.' deleted successfully');
     }
